improvement, update test set
