+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/002_1.lua
@@ -6,7 +6,6 @@ local common = require("test_scripts/Protocol/commonProtocol")
 
 --[[ Local Variables ]]
 local paramsToExclude = { "model", "modelYear", "trim", "systemHardwareVersion" }
-local removeCreatedSession = true
 
 --[[ Local Functions ]]
 local function setHMICap(pParamsToExclude)
@@ -28,24 +27,12 @@ for _, parameter in common.spairs(paramsToExclude) do
   common.Title("Preconditions")
   common.Step("Clean environment", common.preconditions)
   local hmiCap = setHMICap({ parameter })
-  common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { hmiCap })
+  common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { hmiCap })
 
   common.Title("Test")
   common.Step("Vehicle type data without " .. parameter .. " in StartServiceAck", common.startRpcService,
     { common.getRpcServiceAckParams(hmiCap) })
 
   common.Title("Postconditions")
-  common.Step("Stop SDL", common.postconditions, { removeCreatedSession })
+  common.Step("Stop SDL", common.postconditions)
 end
-
-common.Title("Test with excluding all not mandatory parameters")
-common.Title("Preconditions")
-common.Step("Clean environment", common.preconditions)
-common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { setHMICap(paramsToExclude) })
-
-common.Title("Test")
-common.Step("Vehicle type data without all not mandatory params in StartServiceAck", common.startRpcService,
-  { common.getRpcServiceAckParams(setHMICap(paramsToExclude)) })
-
-common.Title("Postconditions")
-common.Step("Stop SDL", common.postconditions)
