+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/002_2.lua
@@ -5,7 +5,6 @@
 local common = require("test_scripts/Protocol/commonProtocol")
 
 --[[ Local Variables ]]
-local removeCreatedSession = true
 local paramsToExclude = { "model", "modelYear", "trim", "systemHardwareVersion" }
 
 --[[ Local Functions ]]
@@ -22,42 +21,16 @@ local function setHMICap(pParamsToExclude)
   return out
 end
 
-local function registerApp(pParamsToExclude)
-  local vehicleData  = common.cloneTable(common.vehicleTypeInfoParams.default)
-  for _, value in pairs(pParamsToExclude) do
-    for key in pairs(vehicleData) do
-      if key == value then vehicleData[key] = nil end
-    end
-  end
-  common.registerAppEx(vehicleData)
-end
-
 --[[ Scenario ]]
-for _, parameter in common.spairs(paramsToExclude) do
-  common.Title("Test with excluding " .. parameter .. " parameter")
-  common.Title("Preconditions")
-  common.Step("Clean environment", common.preconditions)
-  local hmiCap = setHMICap({ parameter })
-  common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { hmiCap })
-
-  common.Title("Test")
-  common.Step("Vehicle type data without " .. parameter .. " in StartServiceAck", common.startRpcService,
-    { common.getRpcServiceAckParams(hmiCap) })
-  common.Step("Vehicle type data without " .. parameter .. " in RAI response", registerApp, { { parameter } })
-
-  common.Title("Postconditions")
-  common.Step("Stop SDL", common.postconditions, { removeCreatedSession })
-end
-
 common.Title("Test with excluding all not mandatory parameters")
 common.Title("Preconditions")
 common.Step("Clean environment", common.preconditions)
-common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { setHMICap(paramsToExclude) })
+local hmiCap = setHMICap(paramsToExclude)
+common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { hmiCap })
 
 common.Title("Test")
 common.Step("Vehicle type data without all not mandatory params in StartServiceAck", common.startRpcService,
-  { common.getRpcServiceAckParams(setHMICap(paramsToExclude)) })
-common.Step("Vehicle type data without all not mandatory params in RAI response", registerApp, { paramsToExclude })
+  { common.getRpcServiceAckParams(hmiCap) })
 
 common.Title("Postconditions")
 common.Step("Stop SDL", common.postconditions)
