+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/006.lua
@@ -18,7 +18,7 @@ end
 --[[ Scenario ]]
 common.Title("Preconditions")
 common.Step("Clean environment", common.preconditions)
-common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { hmiCap })
+common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { hmiCap })
 
 common.Title("Test")
 common.Step("Start RPC Service, Vehicle type data in StartServiceAck", common.startRpcService, { rpcServiceAckParams })
