+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_2.lua
@@ -44,7 +44,7 @@ end
 local function startNoResponseGetVehicleType()
   local hmiCap = common.setHMIcap(common.vehicleTypeInfoParams.custom)
   hmiCap.VehicleInfo.GetVehicleType = nil
-  common.startWithCustomCap(hmiCap)
+  common.start(hmiCap)
 end
