+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_4.lua
@@ -48,7 +48,7 @@ end
 common.Title("Preconditions")
 common.Step("Clean environment", common.preconditions)
 common.Step("Update HMI capabilities", updateHMICapabilitiesFile)
-common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { hmiCap })
+common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { hmiCap })
 
 common.Title("Test")
 common.Step("Start RPC Service, Vehicle type data in StartServiceAck",
