+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/012_1.lua
@@ -31,7 +31,7 @@ for tc, data in common.spairs(tcs) do
   local hmiCap = setHmiCap(data)
 
   common.Title("Test")
-  common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { hmiCap })
+  common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { hmiCap })
   common.Step("Start RPC Service, Vehicle type data in StartServiceAck", common.startRpcService,
     { rpcServiceAckParams })
