+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/012_2.lua
@@ -3,13 +3,6 @@
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local common = require("test_scripts/Protocol/commonProtocol")
-local test = require("user_modules/dummy_connecttest")
-
---[[ Overridden Functions ]]
-local initHMI_onReady_Orig = test.initHMI_onReady
-function test:initHMI_onReady(hmi_table)
-  return initHMI_onReady_Orig(self, hmi_table, false)
-end
 
 --[[ Local Variables ]]
 local tcs = {
@@ -57,12 +50,12 @@ for tc, data in common.spairs(tcs) do
   common.Title("TC[" .. string.format("%03d", tc) .. "]")
   common.Title("Preconditions")
   common.Step("Clean environment", common.preconditions)
-  common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { defaultHmiCap })
+  common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { defaultHmiCap, common.isCacheUsed })
   common.Step("Ignition off", common.ignitionOff)
   local customHmiCap = setHmiCap(data, common.vehicleTypeInfoParams.custom)
 
   common.Title("Test")
-  common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { customHmiCap })
+  common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { customHmiCap, common.isCacheUsed })
   common.Step("Start RPC Service, Vehicle type data in StartServiceAck", common.startRpcService,
     { getRpcServiceAckParams(vehicleTypeInfoParams) })
