+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/013.lua
@@ -23,7 +23,7 @@ end
 --[[ Scenario ]]
 common.Title("Preconditions")
 common.Step("Clean environment", common.preconditions)
-common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { hmiCap })
+common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { hmiCap })
 
 common.Title("Test")
 common.Step("Vehicle type data in RAI", registerApp, { common.vehicleTypeInfoParams.default })
