+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/014_1.lua
@@ -3,29 +3,15 @@
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local common = require("test_scripts/Protocol/commonProtocol")
-local test = require("user_modules/dummy_connecttest")
-
---[[ Overridden Functions ]]
-local initHMI_onReady_Orig = test.initHMI_onReady
-function test:initHMI_onReady(hmi_table)
-  return initHMI_onReady_Orig(self, hmi_table, false)
-end
 
 --[[ Local Variables ]]
-local initialVehicleTypeParams = {
-  make = "OEM2",
-  model = "Ranger",
-  modelYear = "2021",
-  trim = "Base"
-}
-
 local vehicleTypeInfoParams = {
-  make = initialVehicleTypeParams.make,
-  model = initialVehicleTypeParams.model,
-  modelYear = initialVehicleTypeParams.modelYear,
-  trim = initialVehicleTypeParams.trim,
-  ccpu_version = common.vehicleTypeInfoParams.custom.ccpu_version,
-  systemHardwareVersion = common.vehicleTypeInfoParams.custom.systemHardwareVersion
+  make = common.vehicleTypeInfoParams.custom.make,
+  model = common.vehicleTypeInfoParams.custom.model,
+  modelYear = common.vehicleTypeInfoParams.custom.modelYear,
+  trim = common.vehicleTypeInfoParams.custom.trim,
+  ccpu_version = common.vehicleTypeInfoParams.default.ccpu_version,
+  systemHardwareVersion = common.vehicleTypeInfoParams.default.systemHardwareVersion
 }
 
 local defaultHmiCap = common.setHMIcap(common.vehicleTypeInfoParams.default)
@@ -48,34 +34,28 @@ local function getRpcServiceAckParams(pVehicleTypeInfoParams)
   return ackParams
 end
 
-local function startNoResponseGetVehicleType(pHmiCap)
-  local hmiCap = common.setHMIcap(pHmiCap)
-  hmiCap.VehicleInfo.GetVehicleType = nil
-  common.startWithCustomCap(hmiCap)
-end
-
-local function updateHMICapabilitiesFile(pVehicleTypeParams)
-  local hmiCapTbl = common.getHMICapabilitiesFromFile()
-  hmiCapTbl.VehicleInfo.vehicleType.make = pVehicleTypeParams.make
-  hmiCapTbl.VehicleInfo.vehicleType.model = pVehicleTypeParams.model
-  hmiCapTbl.VehicleInfo.vehicleType.modelYear = pVehicleTypeParams.modelYear
-  hmiCapTbl.VehicleInfo.vehicleType.trim = pVehicleTypeParams.trim
-  common.setHMICapabilitiesToFile(hmiCapTbl)
+local function startErrorResponseGetSystemInfo()
+  local hmiCap = common.setHMIcap(common.vehicleTypeInfoParams.custom)
+  hmiCap.BasicCommunication.GetSystemInfo = nil
+  common.start(hmiCap, common.isCacheUsed)
+  common.getHMIConnection():ExpectRequest("BasicCommunication.GetSystemInfo")
+  :Do(function(_, data)
+    common.getHMIConnection():SendError(data.id, data.method, "GENERIC_ERROR", "info message")
+  end)
 end
 
 --[[ Scenario ]]
 common.Title("Preconditions")
 common.Step("Clean environment", common.preconditions)
-common.Step("Update HMI capabilities", updateHMICapabilitiesFile, { initialVehicleTypeParams })
-common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { defaultHmiCap })
+common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { defaultHmiCap, common.isCacheUsed })
 common.Step("Ignition off", common.ignitionOff)
 
 common.Title("Test")
-common.Step("Start SDL, HMI, connect Mobile, start Session",
-  startNoResponseGetVehicleType, { common.vehicleTypeInfoParams.custom })
+common.Step("Start SDL, HMI sends GetSystemInfo(GENERIC_ERROR) response", startErrorResponseGetSystemInfo )
 common.Step("Start RPC Service, Vehicle type data in StartServiceAck",
   common.startRpcService, { getRpcServiceAckParams(vehicleTypeInfoParams) })
 common.Step("Vehicle type data in RAI", common.registerAppEx, { vehicleTypeInfoParams })
 
 common.Title("Postconditions")
 common.Step("Stop SDL", common.postconditions)
+
