+++ test_scripts/Protocol/Enable_OEM_exclusive_apps_support/017.lua
@@ -6,42 +6,25 @@ local common = require("test_scripts/Protocol/commonProtocol")
 
 --[[ Local Variables ]]
 local hmiCap = common.setHMIcap(common.vehicleTypeInfoParams.default)
+local rpcServiceAckParams = common.getRpcServiceAckParams(hmiCap)
 
-local videoServiceParams = {
+local rpcServiceParams = {
   reqParams = {
-    height        = { type = common.bsonType.INT32,  value = 350 },
-    width         = { type = common.bsonType.INT32,  value = 800 },
-    videoProtocol = { type = common.bsonType.STRING, value = "RAW" },
-    videoCodec    = { type = common.bsonType.STRING, value = "H264" }
+    protocolVersion = { type = common.bsonType.STRING, value = "5.3.0" }
   }
 }
-
-local audioServiceParams = {
-  reqParams = {
-    mtu = { type = common.bsonType.INT64,  value = 131072 }
-  }
-}
-
---[[ Local Functions ]]
-local function setVideoConfig()
-  common.getHMIConnection():ExpectRequest("Navigation.SetVideoConfig")
-  :Do(function(_, data)
-      common.getHMIConnection():SendResponse(data.id, data.method, "SUCCESS", {})
-    end)
-end
 
 --[[ Scenario ]]
 common.Title("Preconditions")
 common.Step("Clean environment", common.preconditions)
-common.Step("Start SDL, HMI, connect Mobile, start Session", common.startWithCustomCap, { hmiCap })
+common.Step("Init SDL certificates", common.initSDLCertificates, { "./files/Security/client_credential.pem" })
+common.Step("Start SDL, HMI, connect Mobile, start Session", common.start, { hmiCap })
+common.Step("Register App", common.registerApp)
+common.Step("Activate App", common.activateApp)
 
 common.Title("Test")
-common.Step("Register App", common.registerAppUpdatedProtocolVersion)
-common.Step("Activate App", common.activateApp)
-common.Step("Start unprotected Video Service, ACK", common.startServiceUnprotectedACK,
-  { 1, common.serviceType.VIDEO, videoServiceParams.reqParams, videoServiceParams.reqParams, setVideoConfig })
-common.Step("Start unprotected Audio Service, ACK", common.startServiceUnprotectedACK,
-  { 1, common.serviceType.PCM, audioServiceParams.reqParams, audioServiceParams.reqParams })
+common.Step("Switch RPC Service to Protected mode ACK",
+  common.startServiceProtectedACK, { 1, common.serviceType.RPC, rpcServiceParams.reqParams, rpcServiceAckParams })
 
 common.Title("Postconditions")
 common.Step("Stop SDL", common.postconditions)
