+++ test_sets/oem_exclusive.txt
@@ -1,22 +1,18 @@
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/001.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/002.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/003.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/004.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/005.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/006.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_1.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_2.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_3.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/007_4.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/009.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/010.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/011.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/012_1.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/012_2.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/013.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/015.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_1.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/016_2.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/017.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/018.lua
-./test_scripts/Protocol/Enable_OEM_exclusive_apps_support/019_1.lua
+./test_scripts/Protocol/NACKReason/001_Second_Protected_Video_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/002_Second_Protected_Audio_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/003_Second_Protected_RPC_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/004_Invalid_Certificate_NACK.lua
+./test_scripts/Protocol/NACKReason/005_Wrong_app_type_Video_Audio_Services_NACK.lua
+./test_scripts/Protocol/NACKReason/006_Wrong_HMI_level_Video_Audio_Services_NACK.lua
+./test_scripts/Protocol/NACKReason/007_Force_protected_Video_Audio_Services_NACK.lua
+./test_scripts/Protocol/NACKReason/008_Force_not_protected_Video_Audio_Services_NACK.lua
+./test_scripts/Protocol/NACKReason/009_PTU_failed_Video_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/010_PTU_failed_Audio_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/011_PTU_failed_RPC_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/012_Invalid_time_Video_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/013_Invalid_time_Audio_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/014_Invalid_time_RPC_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/015_Second_unprotected_Video_Audio_RPC_Services_NACK.lua
+./test_scripts/Protocol/NACKReason/016_SetVideoConfig_no_response_Video_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/017_SetVideoConfig_error_response_Video_Service_NACK.lua
+./test_scripts/Protocol/NACKReason/018_SetVideoConfig_error_response_with_rejected_params_Video_Service_NACK.lua
